Add option for custom template
The bootstrap-datepicker library has an (undocumented?) option to replace the HTML template that it renders. I prefer to remove the ranges HTML altogether, so I needed to pass a different template. I can understand if you prefer not to merge this, since it may break and opens some security risks, but it does work.

Usage is simple: pass `customTemplate` to the component—maybe a computed property that emits a string of HTML.